Title: 길이 단위 환산하기, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240717/길이 단위 환산하기/convert-length-units.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    float temp;
+    double temp;
     cin>> temp;
     cout.fixed;
     cout.precision(1);
